# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend cho ứng dụng SmartSchedule - Lịch hẹn thông minh",
        "title": "SmartSchedule API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@smartschedule.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:7070",
    "basePath": "/api/v1",
    "paths": {
...
        "/private/invitations/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trả về danh sách các lời mời sự kiện đang chờ người dùng xác nhận",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Lấy lời mời chờ duyệt",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.InvitationResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/invitations/pending/count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trả về số lượng lời mời đang chờ xác nhận",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Đếm số lời mời chờ duyệt",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/invitations/{id}/accept": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Xác nhận tham gia sự kiện và thêm vào lịch Google",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Chấp nhận lời mời",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InvitationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/invitations/{id}/decline": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Từ chối tham gia sự kiện",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Từ chối lời mời",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.EventDataDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "meeting_link": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.InvitationResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "event_data": {
                    "$ref": "#/definitions/dto.EventDataDTO"
                },
                "event_google_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errors.ErrorCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                2006,
                2007,
                2008,
                2009,
                2010,
                3011,
                3012,
                3013,
                3014,
                3015,
                3016,
                3017,
                3018,
                3019,
                3020,
                3021,
                4022,
                4023,
                4024,
                4025,
                5026,
                5027,
                5028,
                5029,
                6030,
                6031,
                6032,
                6033,
                7034,
                7035
            ],
            "x-enum-varnames": [
                "ErrInvalidCredentials",
                "ErrTokenExpired",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrMissingAuthorizationHeader",
                "ErrInvalidTokenFormat",
                "ErrInvalidInput",
                "ErrInvalidEmail",
                "ErrInvalidPassword",
                "ErrInvalidFormat",
                "ErrInvalidRequestData",
                "ErrNotFound",
                "ErrAlreadyExists",
                "ErrResourceLocked",
                "ErrResourceExpired",
                "ErrCreateFailed",
                "ErrGetFailed",
                "ErrInvalidSlug",
                "ErrDuplicateSlug",
                "ErrInvalidParent",
                "ErrUpdateFailed",
                "ErrDeleteFailed",
                "ErrDatabase",
                "ErrDatabaseTimeout",
                "ErrUniqueViolation",
                "ErrForeignKey",
                "ErrBusinessRule",
                "ErrInvalidState",
                "ErrLimitExceeded",
                "ErrOperationFailed",
                "ErrInternalServer",
                "ErrConfiguration",
                "ErrThirdParty",
                "ErrNetwork",
                "ErrUploadToR2",
                "ErrSaveStorageDb"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}